Add test_mode, test_round
